+++ .gitlab-ci.yml
@@ -32,7 +32,6 @@
     RUN_IDENTIFIER: "${CI_PIPELINE_ID}"
     REPO_NAME: "NeoFOAM-BenchmarkData"
     TARGET_REPO: "github.com/exasim-project/${REPO_NAME}.git"
-    TARGET_BRANCH: "ci-benchmarks"
 
   before_script:
     - pip3 install --user --break-system-packages xmltodict
