+++ benchmarks/dsl/transportOperator.cpp
@@ -8,40 +8,11 @@
 #include "NeoN/NeoN.hpp"
 #include "benchmarks/catch_main.hpp"
 #include "test/catch2/executorGenerator.hpp"
+#include "benchmarks/finiteVolume/cellCentred/common.hpp"
 
 #include "NeoN/dsl/explicit.hpp"
 #include "NeoN/dsl/implicit.hpp"
 #include "NeoN/dsl/expression.hpp"
-
-template<typename ValueType>
-struct CreateVector
-{
-    std::string name;
-    const NeoN::UnstructuredMesh& mesh;
-    std::int64_t timeIndex = 0;
-    std::int64_t iterationIndex = 0;
-    std::int64_t subCycleIndex = 0;
-
-    NeoN::Document operator()(NeoN::Database& db)
-    {
-        auto bcs = fvcc::createCalculatedBCs<fvcc::VolumeBoundary<ValueType>>(mesh);
-        NeoN::Field<ValueType> domainVector(
-            mesh.exec(),
-            NeoN::Vector<ValueType>(mesh.exec(), mesh.nCells(), NeoN::one<ValueType>()),
-            mesh.boundaryMesh().offset()
-        );
-        fvcc::VolumeField<ValueType> vf(mesh.exec(), name, mesh, domainVector, bcs, db, "", "");
-
-        return NeoN::Document(
-            {{"name", vf.name},
-             {"timeIndex", timeIndex},
-             {"iterationIndex", iterationIndex},
-             {"subCycleIndex", subCycleIndex},
-             {"field", vf}},
-            fvcc::validateVectorDoc
-        );
-    }
-};
 
 TEMPLATE_TEST_CASE("TransportOperator::transport", "[bench]", NeoN::scalar, NeoN::Vec3)
 {
